# New security options, new overloads for AddJWTBearerAuth and sensible defaults