--- a/Src/Library/Config/SecurityOptions.cs
+++ b/Src/Library/Config/SecurityOptions.cs
@@ -6,6 +6,15 @@
 public sealed class SecurityOptions
 {
     /// <summary>
+    /// specify a custom claim type used to identity the name of a user principal. defaults to `name`.
+    /// <para>
+    /// WARNING: do not change the default unless you fully comprehend what you're doing!!!
+    /// </para>
+    /// </summary>
+    public string NameClaimType { internal get; set; }
+        = "name";
+    
+    /// <summary>
     /// specify a custom claim type used to identify permissions of a user principal. defaults to `permission`.
     /// <para>
     /// WARNING: do not change the default unless you fully comprehend what you're doing!!!
@@ -13,4 +22,13 @@ public sealed class SecurityOptions
     /// </summary>
     public string PermissionsClaimType { internal get; set; }
         = "permissions"; //should never change from "permissions" or third party auth providers such as Auth0 won't work.
+    
+    /// <summary>
+    /// specify a custom claim type used to identify roles of a user principal. defaults to `role`.
+    /// <para>
+    /// WARNING: do not change the default unless you fully comprehend what you're doing!!!
+    /// </para>
+    /// </summary>
+    public string RoleClaimType { internal get; set; }
+        = "role";
 }