--- a/Src/Library/ServiceResolver/ServiceResolver.cs
+++ b/Src/Library/ServiceResolver/ServiceResolver.cs
@@ -7,6 +7,7 @@ namespace FastEndpoints;
 sealed class ServiceResolver : IServiceResolver
 {
     readonly ConcurrentDictionary<Type, ObjectFactory> _factoryCache = new();
+    readonly ConcurrentDictionary<Type, object> _singletonCache = new();
     readonly IServiceProvider _rootServiceProvider;
     readonly IHttpContextAccessor _ctxAccessor;
 
@@ -24,13 +25,13 @@ sealed class ServiceResolver : IServiceResolver
 
     public object CreateInstance(Type type, IServiceProvider? serviceProvider = null)
     {
-        var factory = _factoryCache.GetOrAdd(type, t => ActivatorUtilities.CreateFactory(t, Type.EmptyTypes));
+        var factory = _factoryCache.GetOrAdd(type, ActivatorUtilities.CreateFactory(type, Type.EmptyTypes));
 
         return factory(serviceProvider ?? _ctxAccessor?.HttpContext?.RequestServices ?? _rootServiceProvider, null);
     }
 
     public object CreateSingleton(Type type)
-        => ActivatorUtilities.CreateInstance(_rootServiceProvider, type);
+        => _singletonCache.GetOrAdd(type, ActivatorUtilities.GetServiceOrCreateInstance(_rootServiceProvider, type));
 
     public IServiceScope CreateScope()
         => _isUnitTestMode