Fixes #526 

This:
- Adds new overloads for the configuration of JWTBearerAuth, including the option to configure the whole `JWTBearerOptions`
- Sets sensible defaults for the JWTBearerAuth, specifically:
- - No longer maps inbound claims
- - Overrides the Role claim to `role` (`Conf.SecOpts.RoleClaimType`)
- - Overrides the Name claim to `name` (`Conf.SecOpts.NameClaimType`)
- Adds new security options (`Conf.SecOpts`) for the Role and Name claim so that it can be customised back to the crazy SOAP ones by doing `Conf.SecOpts.RoleClaimType = ClaimTypes.Role` for example
- `JWTBearer.CreateToken` is using the `Conf.SecOpts.RoleClaimType`

All of the above allows for smaller tokens to be generated and consumed, without unexpected behaviour, whilst still being fully customisable.

Note that, due to the change in the claim role type this is a "breaking change", though the impact should be minimal and will only affect older tokens.