@rudiv 2023-11-04
While we're at it, super fun bonus - I've updated `JWTBearer.CreateToken()` to use the new `JsonWebTokenHandler` on .NET 8+ (since that's a compile target now too), so less allocations + more performance. These are preliminary benchmarks basically because my computer's deciding to be a bit rubbish and make .NET 8 slower overall. I'll re-run them in a bit.

| Method                | Runtime  | Mean     | Error     | StdDev    | Ratio | RatioSD | Gen0    | Gen1   | Allocated | Alloc Ratio |
|---------------------- |--------- |---------:|----------:|----------:|------:|--------:|--------:|-------:|----------:|------------:|
| FastEndpoints         | .NET 6.0 | 22.61 us |  8.841 us |  5.261 us |  1.00 |    0.00 |  5.1000 |      - |  10.38 KB |        1.00 |
| 'Token (Old Handler)' | .NET 6.0 | 44.14 us |  1.223 us |  0.640 us |  1.98 |    0.39 | 15.2000 |      - |  30.47 KB |        2.93 |
|                       |          |          |           |           |       |         |         |        |           |             |
| FastEndpoints         | .NET 7.0 | 20.40 us | 11.973 us |  7.920 us |  1.00 |    0.00 |  1.6000 |      - |  10.22 KB |        1.00 |
| 'Token (Old Handler)' | .NET 7.0 | 32.68 us |  2.353 us |  1.231 us |  1.63 |    0.52 |  4.9000 | 0.2000 |  29.89 KB |        2.93 |
|                       |          |          |           |           |       |         |         |        |           |             |
| FastEndpoints         | .NET 8.0 | 29.91 us | 21.522 us | 14.235 us |  1.00 |    0.00 |  1.6000 |      - |  10.08 KB |        1.00 |
| 'Token (Old Handler)' | .NET 8.0 | 29.20 us | 13.522 us |  7.072 us |  0.96 |    0.46 |  4.5000 | 0.2000 |  27.35 KB |        2.71 |
| 'Token (New Handler)' | .NET 8.0 | 36.08 us | 26.100 us | 17.264 us |  1.33 |    0.36 |  2.8000 | 0.1000 |  17.53 KB |        1.74 |