@KateyBee 2023-11-09
Merged and pushed build out to Nuget `v5.19.0.10-beta`

Did a bit of cleaning up and there were a couple of ambiguous method warnings in `AuthExtensions.cs`.

Let me know if you see anything wrong.

Thanks for the contribution!